+++ app/src/main/java/com/example/persistencia/Modelos/ProductoLista.kt
@@ -5,6 +5,6 @@ import com.google.firebase.firestore.IgnoreExtraProperties
 @IgnoreExtraProperties
 data class ProductoLista(
     val id: String = "",           // ID del documento = Código del producto
-    val cantidad: Int = 1,
+    val cantidad: Double = 1.0,
     val comprado: Boolean = false
 )
